requirements.txt
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
comm==0.2.3
cryptography==46.0.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
